moved player structure to the object source file
