+++ src/physics.rs
@@ -1,7 +1,6 @@
-use crate::player::Player;
-use crate::object::Object;
+use crate::object;
 
-pub fn activate_collision(player: &mut Player<'_>, object: &Object<'_>) {
+pub fn activate_collision(player: &mut object::Player<'_>, object: &object::Object<'_>) {
     /* Player and Object collision */
     let vertex_gap: i32 = 2;
     if player.dstrect.y < object.dstrect.y + 50 - vertex_gap && player.dstrect.y + 50 > object.dstrect.y + vertex_gap {
